#163: Reactive extension 파일 생성 > isSelected 프로퍼티 추가
